feat(rec): store Duration header

+++ pkg/asciicast/asciicast.go
@@ -15,8 +15,8 @@ type Header struct {
 	Width         int     `json:"width"`
 	Height        int     `json:"height"`
 	Timestamp     int64   `json:"timestamp,omitempty"`
-	Duration      float32 `json:"duration,omitempty"`
-	IdleTimeLimit float32 `json:"idle_time_limit,omitempty"`
+	Duration      float64 `json:"duration,omitempty"`
+	IdleTimeLimit float64 `json:"idle_time_limit,omitempty"`
 	Command       string  `json:"command,omitempty"`
 	Title         string  `json:"string,omitempty"`
 	Env           struct {
